@@ -12,41 +12,56 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^10.0.2",
-    "@angular/cdk": "^10.0.1",
-    "@angular/common": "^10.0.2",
-    "@angular/compiler": "^10.0.2",
-    "@angular/core": "^10.0.2",
-    "@angular/forms": "~10.0.2",
-    "@angular/platform-browser": "~10.0.2",
-    "@angular/platform-browser-dynamic": "~10.0.2",
-    "@angular/router": "~10.0.2",
+    "@angular/animations": "12.1.2",
+    "@angular/cdk": "12.1.2",
+    "@angular/common": "12.1.2",
+    "@angular/compiler": "12.1.2",
+    "@angular/core": "12.1.2",
+    "@angular/forms": "12.1.2",
+    "@angular/platform-browser": "12.1.2",
+    "@angular/platform-browser-dynamic": "12.1.2",
+    "@angular/router": "12.1.2",
+    "@ngtools/webpack": "12.1.2",
     "@aspnet/signalr": "^1.1.4",
-    "@ngtools/webpack": "^10.0.1",
+    "rxjs": "~6.6.0",
+    "tslib": "2.2.0",
+    "zone.js": "~0.11.4",
     "assert": "^2.0.0",
     "core-js": "^3.6.1",
     "crossfade": "^1.0.0",
     "jasmine-core": "^3.5.0",
     "jasmine-spec-reporter": "^5.0.2",
-    "rxjs": "~6.6.0",
     "subsink": "^1.0.0",
-    "tslib": "^1.10.0",
     "typed-web-workers": "^3.0.0-next",
     "web-animations-js": "^2.3.2",
-    "zone.js": "~0.10.3"
+    "socket.io": ">=2.4.0",
+    "socket.io-parser": ">=3.3.2",
+    "glob-parent": ">=5.1.2",
+    "dns-packet": ">=1.3.2",
+    "browserslist": ">=4.16.5",
+    "hosted-git-info": ">=2.8.9",
+    "lodash": ">=4.17.21",
+    "url-parse": ">=1.5.0",
+    "handlebars": ">=4.7.7",
+    "xmlhttprequest-ssl": ">=1.6.2",
+    "ssri": ">=6.0.2",
+    "sockjs": ">=0.3.20",
+    "y18n": ">=4.0.1",
+    "elliptic": ">=6.5.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1000.1",
-    "@angular/cli": "^10.0.1",
-    "@angular/compiler-cli": "^10.0.2",
-    "@angular/language-service": "~10.0.2",
+    "@angular-devkit/build-angular": "12.1.2",
+    "@angular/cli": "12.1.2",
+    "@angular/compiler-cli": "12.1.2",
+    "@angular/language-service": "12.1.2",
     "@types/assert": "^1.5.1",
     "@types/jasmine": "^3.5.11",
     "@types/jasminewd2": "^2.0.8",
     "@types/karma": "^5.0.0",
     "@types/node": "^14.0.19",
     "@types/signalr": "^2.2.35",
-    "karma": "^5.1.0",
+    "typescript": "4.2",
+    "karma": "^6.3.4",
     "karma-chrome-launcher": "^3.1.0",
     "karma-coverage-istanbul-reporter": "^3.0.3",
     "karma-jasmine": "^3.3.1",
@@ -55,6 +70,6 @@
     "protractor": "^7.0.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.15.0",
-    "typescript": "^3.9.6"
+    "postcss": ">=7.0.36"
   }
 }
